HealthcarePatientAPIPython/requirements.txt
Flask==2.3.2
pymongo==4.1.1
python-consul==1.1.0
